pkg/archive: various cleanup
* split out platform specific logic
* acknowledge error in Lstat
* close file handler in tests
