+++ pkg/archive/seenfiles.go
@@ -1,0 +1,4 @@
+package archive
+
+// for hardlink mapping
+type SeenFiles map[uint64]string
